Add ToolPatch for handling patching functionality
Integrate ToolPatch into tools module and ensure correct ticket handling
