+++ src/tools/mod.rs
@@ -4,10 +4,11 @@ pub mod tools_execute;
 mod tool_ast_definition;
 mod tool_search;
 mod tool_ast_reference;
-pub mod patch;
+pub mod tool_patch_aux;
 mod tool_web;
 mod tool_tree;
 mod tool_knowledge;
 mod tool_relevant_files;
 mod tool_cat;
 mod tool_locate_search;
+pub mod tool_patch;
