--- src/tools/patch/mod.rs
@@ -1,8 +1,0 @@
-mod diff_formats;
-mod chat_interaction;
-pub mod tool_patch;
-mod ast_interaction;
-mod unified_diff_format;
-pub mod tickets;
-mod ticket_to_chunks;
-mod patch_utils;
